detection/requirements.txt
openvino==2024.6.0
onnxruntime-openvino==1.20.0
opencv-python-headless==4.10.0.84
requests
numpy
Pillow
huggingface-hub
tqdm